+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to GOBL will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/). See also the [GOBL versions](https://docs.gobl.org/overview/versions) documentation site for more details.
 
-## [XXXXXX] - XXXX-XX-XX
+## [v0.77.0] - 2024-05-16
+
+Fixing important bugs with some tax regimes and tax identity code validation, along with a new 'series' property for correction options.
 
 ### Added